fix: unblock the loop -- it could never complete an iteration
Two independent bugs, both found by a --skip-llm dry run, each of which alone
prevented any iteration from ever reaching N60. Both pre-existing runs under
runs/ end in SCOPE_VIOLATION; this is why.

1. N13 rejected every iteration, model or no model.

   Elaboration emits gemmini_params.h into the NESTED gemmini-rocc-tests
   submodule (constants.py:GEMMINI_PARAMS_H_REL). SUBMODULES listed only the
   outer generators/gemmini, whose `git status` reports the nested one as a
   single opaque " M software/gemmini-rocc-tests" -- a path outside the
   writable set, indistinguishable from a model edit. Verified in-container:
   the only dirty file is include/gemmini_params.h.

   Fixed by recursing into the nested submodule (per-file paths instead of an
   aggregate), adding --ignore-submodules=all to the per-submodule status so
   pointers do not double-report, and excusing the generated header BY NAME in
   loop.py's harness_paths.

   This makes the scope check STRICTER, not looser. Before, any edit anywhere
   under gemmini-rocc-tests collapsed into one opaque entry. Now each file is
   reported individually and only gemmini_params.h is excused -- asserted
   against check_patch_scope: the header and SparseCraftParams.scala pass, a
   .c file in the same submodule, another header there, and gemmini RTL are
   all still rejected.

2. metrics.parse raised AFTER a full elaborate and an 11-minute simulate.

   The kernel printed 64-bit values with %llu, but the build links newlib-NANO
   (-specs=htif_nano.specs), whose printf has no long-long conversion. The
   cached simulator log shows the literal result:

       SPARSECRAFT cycles=lu
       SPARSECRAFT macs_useful=lu
       SPARSECRAFT EXE_ACTIVE_CYCLE=26703      <- %u, fine

   cycles and macs_useful are both REQUIRED, so every iteration died at the
   very last step, after paying the entire compute cost.

   -u _printf_long_long is the textbook fix and does NOT work here; verified
   under spike, which prints "A llu=lu" with the flag applied. %lu with an
   (unsigned long) cast does work and costs no range, since long is 64 bits on
   RV64 LP64. Applied that instead and backed the no-op flag out.

Verified end to end: the first complete iteration this repo has produced.

  N50 measured cycles=89,986  off-chip=1,253,376B
  N60 ADMIT_FRONT  front=1  niches=1
  diagnosis: bank-conflict bound (conflict_stalls=4.88) -- lever L3 banking
  wall_clock_s: 1089.41

+++ kernels/attn_prefill.c
@@ -4,6 +4,14 @@
 // simulator's HTIF stdout, so anything the harness scores has to be printed
 // here in the exact "SPARSECRAFT <name>=<int>" form metrics.py parses.
 //
+// 64-bit values use %lu / (unsigned long), NEVER %llu. This build links
+// newlib-NANO (-specs=htif_nano.specs in nodes.py), whose printf has no
+// long-long conversion: "%llu" emits the literal text "lu" and the metric
+// parses as nothing, so the iteration dies in metrics.parse AFTER paying
+// for a full elaborate and a ~10 min simulate. -u _printf_long_long does
+// NOT fix it (verified under spike). On RV64 LP64 a long is 64 bits, so
+// %lu costs no range.
+//
 // Counters come from Gemmini's CounterFile. The hardware exposes exactly
 // EIGHT slots (counter_read masks the index with 0x7), and a slot only counts
 // from the moment it is configured -- so all eight are armed before the
@@ -136,13 +144,13 @@ int main(void) {
 
   uint64_t macs = nnz_blocks * (uint64_t)BLOCK_SIZE * BLOCK_SIZE * D_HEAD * 2ull;
 
-  printf("SPARSECRAFT cycles=%llu\n", (unsigned long long)(t1 - t0));
-  printf("SPARSECRAFT macs_useful=%llu\n", (unsigned long long)macs);
-  printf("SPARSECRAFT nnz_blocks=%llu\n", (unsigned long long)nnz_blocks);
+  printf("SPARSECRAFT cycles=%lu\n", (unsigned long)(t1 - t0));
+  printf("SPARSECRAFT macs_useful=%lu\n", (unsigned long)macs);
+  printf("SPARSECRAFT nnz_blocks=%lu\n", (unsigned long)nnz_blocks);
   printf("SPARSECRAFT seq_len=%d\n", SEQ_LEN);
   printf("SPARSECRAFT block_size=%d\n", BLOCK_SIZE);
   printf("SPARSECRAFT d_head=%d\n", D_HEAD);
-  printf("SPARSECRAFT checksum=%lld\n", (long long)checksum);
+  printf("SPARSECRAFT checksum=%ld\n", (long)checksum);
 
   printf("SPARSECRAFT EXE_ACTIVE_CYCLE=%u\n", counter_read(C_EXE_ACTIVE));
   printf("SPARSECRAFT LOAD_DMA_WAIT_CYCLE=%u\n", counter_read(C_LOAD_DMA_WAIT));
